requirements.txt
python-telegram-bot==20.7
anthropic==0.8.1
Pillow==10.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9 